WGBH-DCAMM/requirements.txt
PyPDF2==1.26.0
tabula-py